[glyf] Split Glyph.hh
