+++ src/Makefile.sources
@@ -90,6 +90,7 @@ HB_BASE_sources = \
 	hb-ot-layout-gsub-table.hh \
 	OT/glyf/glyf.hh \
 	OT/glyf/loca.hh \
+	OT/glyf/Glyph.hh \
 	OT/Layout/GSUB/Common.hh \
 	OT/Layout/GSUB/Sequence.hh \
 	OT/Layout/GSUB/SingleSubstFormat1.hh \
